pyproject.toml
[project]
name = "spatial-knn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.3",
    "ipykernel>=7.2.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow",
    "numpy>=2",
    "ruff>=0.15.7",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.48",
    "duckdb>=1.5.0",
    "sedonadb>=0.3.0",
    "geoarrow-pyarrow>=0.2.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "tabulate>=0.10.0",
]
